+++ overlays/default.nix
@@ -1,4 +1,11 @@
 [
   (import ./onnxruntime.nix)
   (import ./libqmi.nix)
+  (import ./gom.nix)
+  (import ./arcmenu.nix)
+  (import ./gupnp-av.nix)
+  (import ./gexiv2.nix)
+  (import ./git.nix)
+  (import ./libglycin.nix)
+  (import ./pysmbc.nix)
 ]